backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
pydantic==2.6.4
pydantic-settings==2.2.1
aiofiles==23.2.1
python-multipart==0.0.9
pypdfium2==4.22.0
Pillow==10.2.0
sse-starlette==1.8.2
httpx==0.27.2
requests-unixsocket==0.3.0
docker==7.1.0
html2text==2024.2.26
reportlab==4.0.7
